Added a new incident + recall data
After looking more carefully at the SaferProduct website, it seems like one can export recalls and reports at the same time. I did this for the whole Toys and Children category. I added a screenshot showing the subcategories contained in this data, in particular Strollers and carseats are included. Should we exclude some of these sub-category?